+++ lib/Model/Problem/ProblemRegisterModelV2.dart
@@ -1,62 +1,46 @@
 import 'package:image_picker/image_picker.dart';
-import 'package:ono/Model/TemplateType.dart';
+import 'package:ono/Model/Problem/TemplateType.dart';
 
-class ProblemRegisterModel {
+class ProblemRegisterModelV2 {
   int? problemId;
-  String? problemImageUrl;
-  String? processImageUrl;
-  XFile? solveImage;
+  XFile? problemImage;
   XFile? answerImage;
   String? memo;
   String? reference;
   DateTime? solvedAt;
   int? folderId;
-  TemplateType? templateType;
-  String? analysis;
 
-  ProblemRegisterModel({
+  ProblemRegisterModelV2({
     this.problemId,
-    this.problemImageUrl,
-    this.processImageUrl,
-    this.solveImage,
+    this.problemImage,
     this.answerImage,
     this.memo,
     this.reference,
     this.solvedAt,
     this.folderId,
-    this.templateType,
-    this.analysis,
   });
 
   Map<String, dynamic> toJson() {
     return {
       'problemId' : problemId,
-      'problemImageUrl': problemImageUrl,
-      'processImageUrl' : processImageUrl,
-      'solveImage': solveImage,
+      'problemImage': problemImage,
       'answerImage': answerImage,
       'memo': memo,
       'reference': reference,
       'solvedAt': solvedAt?.subtract(const Duration(hours: 9)).toIso8601String(),
       'folderId' : folderId,
-      'templateType' : templateType,
-      'analysis' : analysis,
     };
   }
 
-  factory ProblemRegisterModel.fromJson(Map<String, dynamic> json) {
-    return ProblemRegisterModel(
+  factory ProblemRegisterModelV2.fromJson(Map<String, dynamic> json) {
+    return ProblemRegisterModelV2(
       problemId: int.parse(json['problemId']),
-      problemImageUrl: json['problemImageUrl'],
-      processImageUrl: json['processImageUrl'],
-      solveImage: json['solveImage'],
+      problemImage: json['problemImage'],
       answerImage: json['answerImage'],
       memo: json['memo'],
       reference: json['reference'],
       solvedAt: DateTime.parse(json['solvedAt']).subtract(const Duration(hours: 9)),
       folderId: int.parse(json['folderId']),
-      templateType: json['templateType'],
-      analysis: json['analysis'],
     );
   }
 }
