+++ lib/Model/Problem/ProblemThumbnailModel.dart
