+++ lib/Provider/FoldersProvider.dart
@@ -6,12 +6,12 @@ import 'package:flutter/material.dart';
 import 'package:ono/Service/Network/HttpService.dart';
 import 'package:ono/GlobalModule/Util/ProblemSorting.dart';
 import 'package:ono/GlobalModule/Util/ReviewHandler.dart';
-import 'package:ono/Model/TemplateType.dart';
+import 'package:ono/Model/Problem/TemplateType.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
 import '../Config/AppConfig.dart';
 import '../Model/Folder/FolderModel.dart';
-import '../Model/ProblemModel.dart';
+import '../Model/Problem/ProblemModel.dart';
 import '../Model/ProblemRegisterModel.dart';
 import '../Model/ProblemRegisterModelV2.dart';
 import 'TokenProvider.dart';
