+++ lib/Provider/PracticeNoteProvider.dart
@@ -2,13 +2,13 @@ import 'dart:convert';
 import 'dart:developer';
 
 import 'package:flutter/material.dart';
-import 'package:ono/Model/PracticeNoteModel.dart';
+import 'package:ono/Model/PracticeNote/PracticeNoteModel.dart';
 import 'package:ono/Model/PracticeNoteRegisterModel.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
 import '../Config/AppConfig.dart';
 import '../Service/Network/HttpService.dart';
-import '../Model/ProblemModel.dart';
+import '../Model/Problem/ProblemModel.dart';
 import 'TokenProvider.dart';
 
 class ProblemPracticeProvider with ChangeNotifier{
