+++ lib/Screen/Folder/DirectoryScreen.dart
@@ -9,7 +9,7 @@ import 'package:ono/GlobalModule/Theme/NoteIconHandler.dart';
 import 'package:ono/GlobalModule/Dialog/SnackBarDialog.dart';
 import 'package:ono/GlobalModule/Util/UrlLauncher.dart';
 import 'package:ono/Model/Folder/FolderModel.dart';
-import 'package:ono/Model/LoginStatus.dart';
+import 'package:ono/Model/Common/LoginStatus.dart';
 import 'package:ono/Model/ProblemRegisterModel.dart';
 import 'package:ono/Provider/FoldersProvider.dart';
 import 'package:provider/provider.dart';
@@ -17,10 +17,10 @@ import '../../GlobalModule/Image/DisplayImage.dart';
 import '../../GlobalModule/Text/StandardText.dart';
 import '../../GlobalModule/Theme/ThemeHandler.dart';
 import '../../GlobalModule/Dialog/FolderSelectionDialog.dart';
-import '../../Model/TemplateType.dart';
+import '../../Model/Problem/TemplateType.dart';
 import '../../Provider/ScreenIndexProvider.dart';
 import '../../Service/ScreenUtil/DirectoryScreenService.dart';
-import '../../Model/ProblemModel.dart';
+import '../../Model/Problem/ProblemModel.dart';
 import '../../Provider/UserProvider.dart';
 import '../UserGuideScreen.dart';
