+++ lib/Screen/ProblemDetail/ProblemDetailScreen.dart
@@ -11,8 +11,8 @@ import '../../GlobalModule/Theme/ThemeHandler.dart';
 import '../../GlobalModule/Dialog/FolderSelectionDialog.dart';
 import '../../GlobalModule/Util/FolderNavigationButtons.dart';
 import '../PracticeNote/PracticeNavigationButtons.dart';
-import '../../Model/ProblemModel.dart';
-import '../../Model/TemplateType.dart';
+import '../../Model/Problem/ProblemModel.dart';
+import '../../Model/Problem/TemplateType.dart';
 import '../../Provider/FoldersProvider.dart';
 import '../../Service/ScreenUtil/ProblemDetailScreenService.dart';
 import '../ProblemShare/AnswerShareScreen.dart';
