+++ lib/Screen/ProblemDetail/Template/SpecialProblemDetailTemplate.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../../../GlobalModule/Theme/ThemeHandler.dart';
-import '../../../Model/ProblemModel.dart';
+import '../../../Model/Problem/ProblemModel.dart';
 import '../ProblemDetailScreenWidget.dart';
 
 class SpecialProblemDetailTemplate extends StatelessWidget {
