+++ lib/Screen/ProblemRegister/ProblemRegisterTemplate.dart
@@ -13,9 +13,9 @@ import '../../GlobalModule/Text/StandardText.dart';
 import '../../GlobalModule/Theme/ThemeHandler.dart';
 import '../../GlobalModule/Dialog/FolderSelectionDialog.dart';
 import '../../GlobalModule/Util/LatexTextHandler.dart';
-import '../../Model/ProblemModel.dart';
+import '../../Model/Problem/ProblemModel.dart';
 import '../../Model/ProblemRegisterModel.dart';
-import '../../Model/TemplateType.dart';
+import '../../Model/Problem/TemplateType.dart';
 import '../../Provider/FoldersProvider.dart';
 import '../../Provider/ScreenIndexProvider.dart';
 import '../../Service/ScreenUtil/ProblemRegisterScreenService.dart';
